refactor(subsonic): extract helpers to own module
- Moved the main subsonic functions from sharedFunctions towards the new individual utility file.
- Refactored the subsonic commands towards using the new module.
